Fix: Rename segmented.tsx
